Linter compliance again
