+++ know/rdf/parser/abstract_parser.py
@@ -1,10 +1,20 @@
+
 from abc import ABC, abstractmethod
+from enum import Enum
+class EntityType(Enum):
+    """ The enum that holds entity types for easy access """
+    #Add more types later
+    PERSON = 1
+    WORK_OF_ART = 2
+    BOOK = 3
+    PAINTING = 4
 
 class AbstractParser(ABC):
     """ The base class for implementing a parser for a particular source"""
     @abstractmethod
-    def __init__(self):
-        pass
+    def __init__(self, uri: str):
+        """ Retrieves entity_id and rdf_uri if necessary """
+        self.uri = uri
 
     @abstractmethod
     def parse(self) -> dict:
