+++ know/rdf/parser/viaf_parser.py
@@ -1,15 +1,19 @@
-from .abstract_parser import AbstractParser
-
+# pylint: disable=no-self-use,unused-import,import-error
+import re
+from abstract_parser import AbstractParser, EntityType
 class ViafParser(AbstractParser):
     """parses viaf sources"""
 
-    def __init__(self, rdf_uri):
-        self.rdf_uri = rdf_uri
+    def __init__(self, uri: str):
+        super().__init__(uri)
+        self.entity_id = re.search('[0-9]+', self.uri)
+        if self.entity_id is None:
+            raise Exception("Invalid VIAF URI")
+        self.rdf_uri = "placeholder"
+        # need to construct rdf_uri from uri
 
     def parse(self) -> dict:
         """call this method to start parsing"""
-        entity_type = self.get_entity_type()
-        #TODO call correct parsing function based on entity type
         return dict()
 
     def get_entity_type(self) -> str:
@@ -17,7 +21,9 @@ class ViafParser(AbstractParser):
         return ""
 
     def parse_person(self) -> dict:
+        """ Parses a person entity type """
         return dict()
 
     def parse_work_of_art(self) -> dict:
+        """ Parses a work of art entity type """
         return dict()
